update para por na condicao do exercicio

+++ puraFuncaoEuler.cpp
@@ -2,8 +2,8 @@
 
 using namespace std;
 
-int fatorial(int n){
-    int f = 1;
+double fatorial(int n){
+    double f = 1;
     for(int i = 1; i<=n; i++){
         f*=i;
     }
